+++ tests/test_cli.py
@@ -7,6 +7,7 @@ import os
 import time
 from pprint import pprint
 import logging
+import importlib
 
 import pytest
 
@@ -20,6 +21,7 @@ import paasify.errors as error
 from paasify.common import get_paasify_pkg_dir
 
 log = logging.getLogger()
+common = importlib.import_module("common_lib")
 
 # Test cli
 # ------------------------
@@ -40,13 +42,13 @@ PRJ_VAR_MERGE = cwd + "/tests/examples/var_merge"
 PRJ_NONE = cwd
 
 
-
 def test_cli_docker_is_running():
     assert os.path.exists("/var/run/docker.sock"), "Docker does not seems to run"
 
 
 # Test things that should fail
 #######################################
+
 
 def test_cli_info_without_project():
     opts = ["-vvvvv", "--config", cwd + "/tests/examples", "info"]
@@ -57,13 +59,12 @@ def test_cli_info_without_project():
 
 
 def test_cli_fail_on_outside_project():
-    prj_dir = cwd + "/tests/examples/real_tests"
+    prj_dir = cwd + PRJ_REAL_TESTS
     opts = ["info"]
     result = runner.invoke(cli_app, opts)
     # out = result.stdout_bytes.decode("utf-8")
-    pprint (result.__dict__)
-    assert result.exit_code != 0 # error.ProjectNotFound.rc
-
+    pprint(result.__dict__)
+    assert result.exit_code != 0  # error.ProjectNotFound.rc
 
 
 # Test basic commands
@@ -73,22 +74,24 @@ show_commands = ["explain", "info", "logs", "ls", "ps", "src-ls", "vars"]
 cmd_always = ["document_conf", "help"]
 cmd_always = ["help"]
 
-@pytest.mark.parametrize('cmd', show_commands)
+
+@pytest.mark.parametrize("cmd", show_commands)
 def test_cli_cmd_inside(cmd):
     opts = ["-c", PRJ_REAL_TESTS, cmd]
-    #print ("TESTING", opts)
-    #log.error("TESTING: {opts}")
-    #time.sleep(1)
+    # print ("TESTING", opts)
+    # log.error("TESTING: {opts}")
+    # time.sleep(1)
     result = runner.invoke(cli_app, opts)
 
 
-@pytest.mark.parametrize('cmd', show_commands)
+@pytest.mark.parametrize("cmd", show_commands)
 def test_cli_cmd_fail_outside(cmd):
     opts = ["-c", PRJ_NONE, cmd]
     result = runner.invoke(cli_app, opts)
     assert result.exit_code != 0
 
-@pytest.mark.parametrize('cmd', cmd_always)
+
+@pytest.mark.parametrize("cmd", cmd_always)
 def test_cli_cmd_never_fail(cmd):
     opts = ["-c", PRJ_NONE, cmd]
     result = runner.invoke(cli_app, opts)
@@ -98,16 +101,28 @@ def test_cli_cmd_never_fail(cmd):
 # Test project commands
 #######################################
 
-
 prj = ["real_tests"]
 
-@pytest.mark.parametrize('prj', prj)
-def test_cli_cmd_inside_basics(prj):
-    base = ["-c", os.path.join(EX_DIR, prj)]
 
-    cmds = ["build", "up", "logs", "ps", "down", "recreate", "apply", "down"]
+@pytest.mark.parametrize("prj", prj)
+def test_cli_cmd_inside_basics(prj, data_regression):
+    root_prj = os.path.join(EX_DIR, prj)
+    base = ["-c", root_prj]
+
+    # Check built files
+    runner.invoke(cli_app, base + ["build"])
+    common.recursive_replace(root_prj, root_prj, os.path.relpath(root_prj))
+    results = common.load_yaml_file_hierarchy(root_prj)
+    data_regression.check(results)
+
+    # Loop over commands
+    cmds = ["up", "logs", "ps", "down", "recreate", "apply", "down"]
     for cmd in cmds:
-        #time.sleep(1)
-        #print ("YOLOO", base + [cmd])
+        # time.sleep(1)
+        # print ("YOLOO", base + [cmd])
+        # continue
         result = runner.invoke(cli_app, base + [cmd])
-        assert result.exit_code == 0, f'Failed on running command: {cmd}, got {result}'
+        assert result.exit_code == 0, f"Failed on running command: {cmd}, got {result}"
+
+    # Clean absolute paths
+    common.recursive_replace(root_prj, root_prj, os.path.relpath(root_prj))