+++ CHANGELOG.md
@@ -1,4 +1,28 @@
 
+## 0.1.2 (2023-02-16)
+
+### feat
+
+- add new project files into git by default
+- add support for git on project creation
+- add collection/project banners
+
+### fix
+
+- minor issues in paasify collection
+- gitignore was incomplete
+- messy logo source caused bad rendering
+- unreadable type on b/w themes
+- user material bad rendering
+- file lookup when path is empty
+- extra_vars short form
+- ignored directories in .paasify
+- missing core plugins
+
+### docs
+
+- Update onboarding doc
+
 ## 0.1.1 (2023-02-14)
 
 ### chore